chore(stoneintg-1008): stop referenciing spec.containerImage
The integration service has begun using status.lastPromotedImage
as the source of truth for the GCL. Initially the spec.containerImage
field was used as a fallback until components could migrate to the new
field.  Now that all elligible components have been manually migrated
to the new field, we can remove references to the old field in our code.

This will end the errors caused by the integration and build services
both attempting to write to spec.containerImage.

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>