feat(stoneintg-1237): move scenario validation to admission controller
Moving scenario validation to the admission controller will prevent race
conditions in which a snapshot attempts to test a scenario before it has
been validated.  Additionally, we no longer need the
"IntegrationTestScenarioValid" status condition because all scenarios
that are admitted to the cluster will be implicitly considered valid.
Code surrounding this validation has been removed. Finally, the code for
the removal of the integration test scenario finalizer has been removed
since it is obsolete.

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>