chore(stoneintg-1196): add unit tests for pipelinerun resourceKind
Refactored constants from the tekton package into a subpackage,
`tekton/constants`. This removed a circular depndency that prevented
code to get a ResolutionRequest from being moved to the loader package
where it could be appropriately mocked.  Once the code was moved to the
loader package mocking was added and unit tests were added for all the
new code.  Finally, a new unit test was added for
tekton.NewIntegrationPipelineRun() that verifies that scenarios with the
"pipelinerun" resourceKind can create pipelineruns.

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>