fix(konflux-10792): use auto-release annotation for cel
Since labels can ony contain alphanumeric characters, dashes, and
underscores, this commit updates the auto-release logic to evaluate an
annotation in the releaseplan with the same name as the auto-release
label. If the annotation does not exist, the integration service will
still fall back on the auto-release label (with a true/false value) to
determine whether the snapshot should be auto-released.  This ensures
backwards compatibility

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>